Fix qsort_r().
First problem was that the bit clearing I used in order_roots() was faulty, and
I should just use the utility function I already had. Second issue is that the
calculation for how much space is remaining when deciding between heapify and
order_roots in the insertion loop was wrong. head is the element I am currently
adding to the list, so the remaining space starts at head+1 (or head+leo[0] in
the scaled version).