vasprintf: Copy argument list.
If an argument list is to be used twice, it must be copied. Or rather, once
passed to a function that might call va_arg, an argument list is only good for
calling va_end() on before reuse. So the argument list used for the first call
to vsnprintf() must be a copy of the input list.