Add termios library.
Decided to add it to the fd directory. Which is a bit peripheral for cfgetspeed
and cfsetspeed.

Speaking of, I have decided for now to not support split baud rates. It is an
esoteric feature that makes things more complicated. This is signalled to
applications by setting the output baud rate together with the input baud rate,
as long as the input baud rate is not 0. POSIX leaves it undefined whether they
mean 0 or B0, but Linux splits the difference by defining B0 to 0, so it doesn't
matter.

Arbitrary baud rates are currently not supported, so they just give EINVAL. That
is a "may fail" I am implementing, because the alternative is setting weird
control flags.