Add pthread_mutexattr API.
Thankfully all mutex attributes boil down to simple flags. Except the type,
which I allocated to bits 0 and 1 for that reason. As previously, I am putting
all the getters into the same source file, as they are used rarely, and in the
rare case one is used, most of the others end up used as well.