Add some more trivial functions.
pthread_getcpuclockid() currently returns a poorly disguised kernel TID, and it
has exactly the same lifetime as a kernel TID. That means that taking the
killlock there wouldn't help, because the encoding of the TID escapes the
function. Not sure how to solve this without burdening everything.

phread_getschedparam() has to do two different syscalls to get the two return
values. Why? Because Linux, I suppose.