Add more pthread scheduler functions.
All of these are not-quite-conformant, but at least the unconformance is inline
with all the other libcs. Namely if the used pthread_t describes a zombie
thread, all calls fail with ESRCH. That error condition is not specified in
POSIX in the normative part, and in the informative part only to say
specifically a different reason which is undefined behavior for us (namely use
of a pthread_t after the end of the thread life time).

I saw in musl that it blocks signals during pthread_getschedparam() and figured
it must be because of having to do two syscalls. But no, it is actually because
killlock must be signal safe. pthread_kill() is on the list of signal-safe
interfaces, and it must take the killlock, and so everyone that needs it must
block signals, including pthread_kill().