Add pthread_condattr_*() API.
Thankfully, CLOCK_REALTIME==0, and that is also the default value of the clock
property, as per POSIX.

The spec says I have to avoid CPU-time clocks, but that means that CLOCKFD
clocks ought to be allowed. Might be useful for someone. And if someone gives me
the SGI clock that no longer works, then __timedwait() will give a failure to
that effect. May have to adjust the __timedwait() calls already present to allow
for that possibility.