Add robust list.
In preparation for pthread_mutex_init(): We need to check if robust lists are
supported so we can properly reject an attempt to create robust pshared mutexes
on old kernels.

The robust list is three machine words, of which the first is a pointer to a
lock entry (which itself is a pointer to the next), the second is the (signed)
offset from the start of a lock entry to the lock word, and the third is a
pending lock entry.

I am defining mutexes so that they start with a "next" pointer, which simplifies
the assignments to it. In the end, they make a linked list, terminated with a
pointer to the head of the list. Why Linux hasn't just chosen a NULL terminated
list, like everyone and their dog would have, the gods only know.