Improve pthread_join().
POSIX recommends returning errors for self-joining and trying to join a detached
thread, so I implemented that.

It also specifies that pthread_join() is to be a cancel point. I implement this
by calling pthread_testcancel() before entering the main loop, and calling a
cancel-point version of __futex_wait(). __tl_sync() remains without cancel
point, but when we have exited the loop, the __tl_sync() ought to not take long.