Refactor getdelim().
I have no way of telling whether getc_unlocked() failed for an error or EOF from
the return value. I can only check the error flag afterwards. But of course the
error flag might have already been set before, so I need to clear it first.

No function except clearerr() is supposed to clear the error flag, so I have to
save it and put it back in case of success. Thankfully the file is locked, so
nobody can see me do this.

I also refactored the reallocation stuff, so I have to write it only once.
getdelim() is a weird function in that it can have partial success. My reading
of the spec is that if getdelim() fails, it still returns the most recently
allocated block. That needn't be null-terminated, but the application can still
inspect it as a byte buffer, since it has a length limit.

Maybe I should check what the competition is doing here.