Absorb preceding free blocks in expand_heap().
This was a weakness of the algorithm. It is based on unifying adjacent free
blocks as much as possible, but when expanding the heap it would leave two free
blocks next to each other. That was not ideal. Now, if I can expand the heap in
place, then I will also gobble up all preceding free chunks as much as possible.