Use uniform growth strategy in all functions.
There are a few functions utilizing a dynamic array with exponential backoff.
They now all use the same growth strategy: A minimum allocation of 16 elements
and a growth factor of 1.5 without constant part. This way at least the
consequences are the same everywhere.

I ran a simulation, and with a growth factor of 2, I get a steadily growing free
block at the start of the heap, while the active part keeps moving backwards in
memory. This is not ideal. The lower growth factor keeps it from wandering,
although it does mean more frequent reallocations.