Forbid absolute timezones when elevated.
Of course the whole point of is_tz_acceptable() is to avoid the caller from
specifying whatever file they want. Only the standard search path can be used,
and cannot be escaped, since the substring "../" is outlawed. Worst case, the
user could specify a directory by ending in "..", but then the read call a
little further down just fails.