Add wait*.
Turns out that wait4() is so old, it is above baseline for all architectures. So
I can simplify waitpid().

waitid() is below baseline in at least x86_64, but there is no older call that
does the same thing. In particular the full exit status is not available
anywhere else, unless I were to catch signals. But that would have significant
side effects.