Add pthread_cond_* API.
I am using the probably simplest possible implementation. I make no effort to
save on system calls, or to comply with the spec regarding waking the
highest-priority thread first. (Most implementations don't do that last one,
actually).

My thinking is that those calling pthread_cond_signal() typically expect someone
else to be waiting already. The design is terribly simple: The condition
variable holds a serial number. The waiter reads this serial while under the
mutex, then waits for it to change. pthread_cond_broadcast() increments the
serial number and wakes *all* threads, pthread_cond_signal() increments it and
wakes *one* thread.

Theoretically, it is possible for a thread to miss a wakeup if the serial number
was incremented 4 billion times without waking the thread. I will cross that
particular bridge when I come to it.

Spurious wakeup is possible if a thread is woken up by signal after the serial
was changed. pthread_cond_wait() shall not return EINTR, but nobody said it
can't wake up spuriously.

For non-process shared conditions, theoretically an improvement is possible
where I manage waiters with a linked list, thus avoiding spurious wakeups. But
that is optimization for a later date.