Reduce __SYSCALL_NARGS definition.
The reason I need an argument past the 0 is that it is technically not allowed
to call a variadic macro without variadic argument. But the variadic argument
doesn't have to contain anything.

When calling a zero-arg syscall, the __SYSCALL_NARGS() call expands to

__SYSCALL_NARGS2(nr, 6, 5, 4, 3, 2, 1, 0, )

now. And thus, there is now an empty arg in the variadic slot.