Make pthread_cancel() async-cancel-safe.
pthread_cancel() is supposed to be async-cancel-safe. This means if it is called
on self with async cancellation in effect, cancellation must take place
immediately. This is why we cannot skip sending the signal on self. This way, if
async cancel is in effect, the signal handler can see that while SIGCANCEL is
blocked, and then enter __cancel() with cancellation disabled.