Change meaning of pic_default.
I thought about it more and decided to just go with something similar to
musl again. I am now taking the fact that __PIC__ is defined as signal
that PIC is desired by default. And in that case I will compile
everything with -fPIC explicitly, but only compile a single set of
object files.