Fix build on aarch64.
I had a whole lot of things wrong, but thankfully that was easily fixed.
The fenv stuff was just copied from PPC, and was wrong on aarch64
because of that.

The most significant changes were to memset, however, because I didn't
know that stp needs 8-byte aligned offsets. Since it copies 16-byte
blocks, I go to 16-byte alignment.

There were also a few things wrong in the cpu_arch.h. stlxr must not
have status register and source register overlap, so I needed to model
that.

And I added a dummy sqrtl() implementation. It just reverts to double
precision, and I probably should improve it at some point, but it gets
the implementation going.