Fix vfprintf() blunders.
The shift-left loop didn't work because of two simple oversights. The line to
pad the outputs with zeroes on the right for the precision was also wrong, and
would do the padding exactly the opposite way around.

There were some detail problems calculating the decimal exponent. For the 'f'
format it ended up one too high. For the 'e' format, if trimming zeroes, the
precision ended up one too low, and that revealed another problem: The
outputting of the remaining digits of the first block did not take precision
into account. It is of course possible to output fewer digits than there are in
the first B1B digit.