Add mbrtowc().
The groundwork was already laid with the __utftab. This just implements the spec
and allows restarting. The spec says that in case we return -1, the conversion
state is undefined, but this would make the internal state basically unusable. I
will always reset the state back to initial after verifying len > 0, so that
calling mbrtowc with a null string reliably resets the state, even if an
incomplete character were in there. No clue how applications are supposed to
take advantage of that, however.