Add system().
This is a pretty standard by-the-books sort of system() implementation using
posix_spawn(). One question still open (to me) is what I should do about
cancellation here. system() is not on the list of allowed cancel points, but
neither musl nor the example in POSIX handle the case. Not sure how this works.
I'll just keep it the way it is currently.