Improve destructor ordering.
Destructor ordering is non-trivial. I currently attempt to order them in the
reverse order of them *finishing* their constructors. This allows them (once
dlopen() is defined) to dynamically load other modules and have them count as
dependencies for the purpose of destructor ordering.

However, this opens the door to missing destructors for libraries that have
already had some constructors run when another thread concurrently calls exit().
To forestall that issue, I first block future dlopen() calls from succeeding,
then wait for all in-progress initializations to have finished. This opens
another problem: What if an initializer calls exit()? In that case, there is a
library on the list of loaded modules with ongoing initialization that will
never finish. So I skip libraries that are being initialized by the current
thread, because they will never finish.