Add strcasecmp().
That function is not specified for multibyte ctypes, which is great for me,
since that means I can define it however I will.

I am now decoding both inputs to the current wide character type. If the
decoding fails, I convert the input to a code unit instead, to still treat
different input bytes as different.

In the POSIX locale, mbtowc() never fails with EILSEQ, and high bytes are
converted to code units, which towlower() just doesn't change. So this becomes
an ASCII case-insensitive compare.

But in UTF-8 locale, this is now capable of correctly mapping UTF-8 strings
together. Which is an extension neither musl nor glibc contain.

Rich recently said these functions should just not be used, and maybe my effort
here is misplaced for that reason. Applications indeed cannot expect any
sensible behavior from this function outside of POSIX locale. But I think I
might need a Unicode case-insensitive compare for string collation one of these
days, so now I have that.

I am putting the _l variants in a different file, because they reference
uselocale(), and not everyone who calls strcasecmp() wants that in their
linkage.