Add start of crypt().
I see no reason for overly clever code to differentiate the crypt variants. Just
a jump table for me, please.

I am going to implement most of the variants from the manpage. It does not
matter whether I like the algs in question or not, they are used in existing
shadow files and I need to be compatible. Only I will not implement DES crypt.
In all my years on Linux I have never seen anyone use that dinosaur.