Test for broken EBP in assembler.
On i386, apparently GCC will not allow one to use EBP if the frame pointer is in
use. Thankfully, that appears to be a global setting, and so a simple compile
test can sus out whether I can use EBP normally or not.

In inline assembler, I am not supposed to push anything to stack. Which means
that once I do, I can no longer use any arguments that might be in memory, as I
have no control over what base register they use, and if they use ESP, the
offsets are wrong. So, taking a page out of musl's book, the first thing I do is
push the desired value to the stack.

Inventing something new afterwards, instead of pushing the old value of EBP at
this point, I XCHG it with the top of the stack. So afterwards, the desired
value is in EBP and the old value of EBP is on stack. Therefore, after the
syscall instruction, a single pop is enough to restore EBP to its prior value.