Fix a few logic errors in the sem code.
In particular, don't spuriously fail when some other thread simultaneously
accesses the counter value. There was an issue in sem_trywait() like this. I
think I got sidetracked by the requirement to not block under any circumstance,
but sem_trywait() cannot be made to block indefinitely, except through very
complicated access patterns so that the CAS always fails. That is not a useful
access pattern, so I ignore it.

sem_clockwait() had a missing "continue", so if some other thread managed to
decrement the counter, it would actually start waiting. That is not good.