Only calculate SysV hash if needed.
On Linux, the GNU_HASH section is preferred, and most libraries don't
even come with a SysV hash table. So the SysV hash of the name is needed
only under rare circumstances. Add to that the fact that some mangled
C++ names can be very long, calculating the SysV hash unconditionally
can be a costly mistake.