Fix length calculation for anonymous segments.
The length of the zeroed part at the end is of course the difference
between memsz and filesz, and the difference between pagesz and eoi is
the amount I have already zeroed with memset, so the length of the
anonymous segment is the difference between these two amounts. And then
I need to round up to the next page size.

Previously I had forgotten to subtract the filesz first, thus
overestimating the segment size. That probably won't matter for
practical things, due to the rounding to page size, but now it is
correcter.