Turn a_pause() by default into a no-op.
a_pause() is only needed to relax the CPU in a spin lock loop. The spin
lock is required to synchronize memory, but that's what the a_swap()
already does. On x86, a pause instruction can do that. But otherwise it
is not really needed for anything. Especially not for having a
time-consuming and useless barrier.