Properly dummy out the tlsdesc stuff.
It occurred to me that nothing actually obligates me to turn the dummy
implementation of the tlsdesc functions into actual functions. They only
exist to fulfill linker dependencies in case the compiler isn't smart
enough to recognize the TLSDESC stuff is unused. And on the C level, the
things are declared as constant char arrays.