Improve initial setting of start.
The variable start in match_bracket() records the most recent code point
actually asked for, so that in case of a range expression, we have a start
point. The initial value of this variable is therefore only needed if a range
expression starts before it can be set. It is set at every normal element and at
every collation element or equivalence class. If the bracket starts with the
special cases of ']' or '-', then we can set start simply to the ASCII code
points, since those are ASCII characters, and all allowed MB encodings are ASCII
compatible.

But still one case remains in which the variable remains undefined, and that is
if the bracket starts with a charclass expression, followed by a '-'. That case
is undefined, it is also not very sensible, but I now set start to an
unmatchable value in that case. The variable still retains its most recent value
if the charclass expression is not the first, so that "A[:lower:]-Z" has the
same effect as "[:lower:]A-Z", but the former is not defined, anyway.