Reduce number of relocation in strerror.c.
I am trying to not make this code too much worse for static linking. In dynamic
linking, having a string list creates a a relative relocation for every string.
And every relative relocation needs processing. They can be packed with linker
options, but still have a runtime cost at startup. And the relocations happen
always, even if strerror() is not used. That happens even in implementations
that support lazy linking, BTW.

So instead of having a direct list of error strings for each code, I am now
creating a structure that contains all the error strings, and a list of offsets.
I can then directly look up the error string offset for the error code and
reference the error string.

Since the list of error codes can have a hole in the middle (e.g. code 41 is
often not filled), I am adding 1 to each offset statically, so that 0 remains as
error value. 0 is the value that will be filled into unused spots.

I am also marking the case that the error code is unknown as unlikely, because I
saw in the compiler output that the compiler would look up the "Unknown error"
string unconditionally. But at least clang doesn't care about that annotation
and continues to look up that pointer.