Fix a few logic errors in __time_to_tz().
First of all, getname() didn't null-terminate the name of the timezone. Now,
most time zones are 3 or 4 characters in length, but it still is bad to not do
that termination.

is_posix_form() had the wrong logic. It's supposed to return 1 if there's a name
at the start, followed by an offset. But I made it return 0 in that case. Only
if the string starts with a name and then not an offset would the function
return 1, and that was just wrong.

And finally, do_tzset() might have called strcmp() with a null pointer, because
oldtz might be 0 if we encountered a TZ value of more than 32 bytes and the
allocation failed.