Add flag probing to configure.py.
I have found that lists are passed to functions by reference. This means I can
easily create a function that appends probed flags to lists of flags.

The logic for the flags to probe is taken from musl. Rich Felker has already
gone through the pain of finding these, so I don't have to. With this, I now no
longer let the compiler produce all warnings, but only the ones that interest
me. This gets rid of a few annoying cases. E.g. GCC 12 has a new attribute
"nonstring", to be applied to char arrays to tell the implementation that
missing null-termination is OK there. So I would need to change the code to tell
the compiler that I indeed know what I am doing.

This is C, not python. I know what I am doing by default, and if not, the
compiler is free to blow me up.

I am now probing for the compiler support library. If none is found, none is
added to the command line, but this may cause linker issues. Nothing I can do,
really, if a user has a compiler without support lib installed.