Add all allocation functions to dynamic list.
Now all references to them from the library will be redirected to the PLT, and
thus to any overload defined in a library. I will need to do some work to figure
out partial overloads.

Note that due to the absence of the dynamic linker from the static lib,
overloading of the allocator in static linking is not supported, unless it is a
complete overload (i.e. with calloc() and aligned_alloc()).