Add indirection layer for malloc replacement.
With this, all public allocation functions except aligned_alloc() are now
implemented by calling internal functions, that can also be accessed from other
parts of libc.

aligned_alloc() is a different beast I have to look at separately. It is linked
to the malloc implementation, so I cannot write a "portable" version for
incomplete replacement. At the moment, replacing malloc and free, but not
aligned_alloc(), leads to a broken version of aligned_alloc() that returns
pointers that cannot be freed. Not sure how to deal with this.

To allow incomplete replacement of malloc(), the public calloc() function now
just calls malloc() and then memset(). But it avoids calling memset() on the
whole range, because large mallocs might be satisfied with fresh anonymous
pages, which read as zero, but trap on write access. I am implementing a poor
man's memset() backwards, avoiding zeroing values that are already zero, to
avoid unnecessarily allocating pages. This might make valgrind moo a bit, but
that's what they get for debugging libc. No, seriously, this is a valgrind bug
that needs fixing there.

For the remainder of libc, I checked where to replace the public allocation
functions with the internal ones. Hard criterion for using the internal
interfaces is holding any kind of lock. Hard criterion for using the external
interfaces is having to return pointers that can be passed to free(). This
leaves many functions with no hard pull either way, which I tended to just leave
calling the public functions. I've seen that musl makes the
posix_spawn_file_actions functions use the internal allocator. There is probably
a reason for that, but it is beyond me right now. I leave them using the public
functions.

This also means that ldso now has a working allocator even after the final
relocation. I changed the comments to reflect that, but the logic remains
unchanged. Only, the "malloc_replaced" flag is now completely useless. Maybe it
will gain functionality in future.