Turn base address into a pointer.
I decided in the end that it was easier to deal with the base address as a char
pointer than as a number. I still have to do some casting, but less so now. I am
also changing the meaning of relro_start and relro_end to be the page-aligned
*offsets* into the library, so the base address still has to be added. And I am
moving when the relro pointers get initialized so it happens earlier. But it
can't happen too early, because we still need to skip relro protection for libc
in stage 2.