Add malloc donation.
I add an interface to give memory over to the malloc pool, and use it for memory
that is allocated but would be otherwise wasted.

Memory mapping only occurs at page granularity. So for every map, there is
memory at the start of the first page and at the end of the last page that might
be usable for malloc, and isn't used for the module itself. For malloc, I need
the memory to be readable and writable, so that is what I filter for, and I also
exclude the relro area because of it. GNU ld tends to put the relro area at the
start of the writable segment, which wipes out one avenue for donation, but such
is life.

The donation itself in my malloc implementation just consists of making a simple
memory area and freeing it. But I don't call free() so valgrind doesn't get
confused. I just make a single chunk that covers what part of the space I can
use and bin it.

The memory must of course not be unmapped. As explained in the comment,
libraries successfully loaded on startup are never unloaded. If relocation
fails, we just quit the program. This is different for modules that are
dynamically opened, though, so once I get to dlopen(), I will have to handle
memory donation differently there.

The goal is to be able to use the donated memory for the purposes of the dynamic
linker already, so that loading the dependencies doesn't fail because I cannot
allocate the library descriptor. Nothing is guaranteed here, though. But libc is
definitely loaded by the kernel already, and the main program also might be, so
that might give us memory we can already use.