Add getopt().
Simple POSIX getopt(). I keep track of which options have already been processed
with a counter. Of course it is possible for the application to change optind so
that optchar >= strlen(argv[optind]), but I really try to avoid the issue by
resetting optchar back to 1 when I read end of arguments. This allows conforming
applications to safely reset optind back to 1 once getopt() has returned -1.

Apparently, the GNU extension of setting optind to 0 for a reset is used in the
wild, and I support it for now. I do not support multibyte options. That is a
musl extension that would make everything more complicated. Fewer features means
fewer things to go wrong.

Not a big fan of the huge nested if statement at the end there. I'll see if I
can clean that up. The logic is correct, though.