Add support for PowerPC64's scv instruction.
Oh boy, those benefits!

So, first of all I added a new function __init_sysinfo() to initialize
the sysinfo in an arch-specific way. I decided to do a similar thing for
ppc as for i386 before, namely to have an assembler function for the
syscall instruction. Unlike with i386 before, I also have to have the
second implementation.

This means I just made all PPC64 implementations slower by hiding the
syscall instruction behind an indirect call.

The syscall is then an indirect call to the sysinfo pointer in TLS. This
only works after TLS is set up, though, so the old way with sc must
remain for many uses, including pre-TLS code, the cancel point, clone(),
and sigreturn.

scv is defined to clobber LR, so I need to save it. And experimentally,
gcc does not create a proper stack frame, so I have to do it myself.

The PPC people said the instruction saves 90 cycles over sc. Some of
that advantage is probably eaten by the indirect call and the spill to
memory, so the benefits aren't as clear.