Remove __hwcap and __sysinfo as global variables.
I don't like global variables, and keep looking for alternatives. Sometimes they
are needed for lifetime reasons, and sometimes even for scoping reasons (e.g. if
the API is dictated by standard). But in this case, both __hwcap and __sysinfo
only existed to be initialized moments before being read in __init_tp(), and
then never mattering again after that. So I am removing them.

__hwcap was always just the contents of the AT_HWCAP aux header, so that gets
hardcoded. __init_sysinfo is renamed to __get_sysinfo and now just directly
returns the sysinfo pointer instead of setting a global variable. The additional
parameters are given to __init_from_phdrs(), a functions that is getting
increasingly misnamed.