Integrate yescrypt.
I have copied the source code from the official openwall repo. I took care not
to remove any of the copyright notices or similar comments. And then I went to
work with a hacksaw.

I had to remove the optimized SSE stuff, since in this build environment, I
don't have access to the compiler intrinsics, and I really didn't feel like
adding those to x86_64's cpu_arch.h. Adding access to the compiler intrinsics
would also allow GCC to override some of my include files, which I don't want.

I also made all yescrypt functions static, and removed the unused functions. At
least the statically unused functions.

yescrypt came with its own BSD-licensed sha256 implementation, which I didn't
copy in favor of my own PD and GPL implementation (all the parts of it that are
original to me are GPL. Good luck with that fine-tooth comb.)

I mean, in the end I just implemented publicly available algorithms. And found
out that the two-paragraph summary of an alg on Wikipedia might not be detailled
enough, actually.

I am keeping the test run alive in this one, and am using it to overwrite the
stack and heap buffers. Not that the heap buffers should be visible, since the
memory is allocated with mmap() (and deallocated with munmap()), so any code
that runs later should not be able to see the allocated memory. And the kernel
should always initialize memory it gives to other processes. But no matter,
since I am overwriting the memory anyway. But since I did not understand the
algorithm completely, I have to hope that I am overwriting enough of it. But I
also didn't want to have the test run consume more memory than the main run.

The test vector is the last normal test case from the yescrypt test code. I just
hope that shorter string means less resource consumption, which is what made me
choose this.