Add bcrypt().
Used the OpenBSD code as reference, but worked out the specification
from it first. It is a pretty standard crypt except for the fact that
salt and hash are not separated with a $ character. They are just
juxtaposed.

The use of a cipher algorithm rather than a hash is also new. I have not
implemented decryption because it is not needed for this application.
The block cipher is also used in an interesting fashion: The plaintext
is a constant string, and it is actually the key that is the important
secret, because that is derived from the password.

The whole thing ends up being kinda slow, but oh well. Maybe it needs
more love from me, or maybe it needs profiling to figure out where the
slowdowns are.