Improve bf_crypt.
A simple cost-12 test case was taking 0.4 seconds to execute. I have now
simplified the parsing of bytes into words and pulled it out of the
loop. The salt always has 16 bytes, so no need to account for any
differences. And while the key can have any length, only the first 72
bytes of it are used at most.

This little trick was getting the time down to 0.3 seconds. Since glibc
takes 0.25 seconds on the same test case, I'm thinking this is good
enough.