Add random().
The original formulation of this alg in NetBSD is a bit wordy, so I simplified
what I could. NetBSD also used a different LCG for both seeding and emergency
mode. This implementation I have here is closer to what musl does.

initstate() and setstate() must both serialize the current state of the
function. For that, I use the first word of the state array. Since all of max,
wi, and ri fit in a byte, that is how they are serialized. In theory, I could do
without serializing either of wi or ri since they are always a constant offset
away from each other, but doing so would have no real benefit.

The API is broken in that it requires the state arrays to be char*, but they are
used as arrays of 32-bit integers. However, bionic, musl, and glibc are using
the state array directly, so I suppose it is not a big problem in practice.

Also, I found that I had somehow forgotten setstate() in stdlib.h. Well, it is
added now.