Add first portion of POSIX aio.
I cannot use io_uring for this, since that isn't supported in all kernel
versions, nor for all file types. The simplest way to fulfill all requirements
is to just fire off a thread for each request.

Per spec, some coherency is required: aio_fsync() and aio_write() with O_APPEND
active must be ordered after all preceding writes on the same FD. For this
reason, I need a per-FD queue of requests. This queue can cache whether the FD
is seekable or appendable, so I don't need to get this every time.

The queue is reference counted and will be automatically deallocated once it is
unreferenced. It becomes referenced in submit() (obviously), and dereferenced in
cleanup(), when the worker thread finishes.

The reference counter is increased a second time in submit(), but never in
worker_thread(), because submit() calls unref_queue() at the end to clean up in
case of bad pthread_create() call. So I need to know the refcount was increased
before that happens if the worker could be created. I don't want to use the
argument semaphore for that, since that is an unrelated use, and because I would
have to reorder the calls in the parent to hang onto the queue longer than
necessary.

Another reason for the per-FD queue is that close() must cancel all running AIO
operations. close() is an async-signal-safe function and must therefore not take
locks that may be taken without signals blocked. This is the reason why the AIO
locks all block signals.

A nice side effect of this is that all AIO threads spawn with signals disabled,
and they remain disabled forever.