Add stdbit.h.
I have seen that glibc has added those interfaces. I had hoped GCC would just
add them instead, since it already has the builtins for that, and also I don't
know how useful these interface are in their own functions. But oh well, guess
I'll have to bite the bullet and implement these functions myself.

stdbit.h is a C23 header file and ordinarily, I could just require a C23
compiler when it is included. But since the source language of this library is
C11, it must be usable in C11 for the declarations. Thus the need for the
__unsequenced macro and the stdbool.h include.