Add stdc_leading_zeroes().
I had to add a_clz() to the architectures where it was missing. My current model
is that a_clz() cannot be called with an argument of 0, because that's what x86
requires. Now, PowerPC could deal with that, and ARM maybe as well (I'd have to
check), and so the functions are suboptimal on those architectures. However, I
don't aim for highest performance anyway, and if an application needs that, the
compiler intrinsics should be used.