Add required lio_listio error.
I have to return EINVAL for a bunch of conditions, many of which don't apply
since I don't define AIO_LISTIO_MAX or AIO_MAX. This is a best-effort
implementation; just try what you want and maybe it will work. There are no
guarantees. That is why I cannot give a maximum on the number of entries that I
can deal with at a time in lio_listio(), or a maximum number of AIO operations
at a time. I can do as much as the rest of the system lets me.