Add stdc_leading_ones().
I decided to implement it directly in terms of the cpu.h functions, rather than
calling stdc_leading_zeroes(), so as not to make the code jump too much around.

Again, I am assuming that a_clz() cannot be called with a zero argument, so I am
excluding the thing that would create a zero argument.