Optimize stdc_leading_ones_ui and stdc_trailing_ones_ui.
I realized that I can just add the "am I on 64-bit?" condition to the normal
condition I already have there. It is a compile time condition, so it won't lead
to runtime code, and will cause there to be no conditional at all on 64-bit
systems, where the expression "~(size_t)x" (for unsigned x) cannot be 0.