Add remaining stdbit API.
I needed a population count primitive. It turns out that only PowerPC64
unconditionally has an instruction for that (kind of). At least it is part of
the Power ISA 2.02, which is close to the oldest one still used (although 32-bit
implementations don't follow it).

Oh, goddang it, I just noticed that the instruction is not present in PowerPC
ISA 2.01, which is the baseline. So I'll have to rework this. According to IBM,
the instruction "is only present in POWER5", which I hope means "at least
POWER5".

AArch64 has the CNT instruction, but only if FEAT_CNT is set, and my
preprocessor-fu is too weak to select for that. I'd have to do more research.

x86 has the POPCNT instruction, which was added as part of the bit manipulation
extension at the same time as SSE4, so it is also not part of the baseline.

So in all of these cases, at best I can add these instructions to the lib if
compiling for an ISA level high enough to be able to assume their existance.
Checking hwcap/cpuid is not an option here, because I can then not simply revert
to the portable implementation if the feature is not present.

Anyway, back to the actual stdbit functions. Most of these are variations on
a_clz(), but requiring some special casing for various arguments. I treat
a_clz() and a_ctz() as undefined if the argument is zero, so that had to be
avoided.

There are a few instances of me using "1 << (c - 1) << 1" instead of "1 << c". I
do that when c is at least one and at most equal to the width of the left type,
because in that case, the latter instruction would be undefined.

Defining the functions for small types is not always straightforward, as C only
supports zero-extension and sign-extension, but a few of these functions would
require 1-extension (especially the "count leading ones" type functions). I
solve the problem by instead shifting the input value to the high part of a
size_t.

I decided to implement all functions in separate files as there is no reason to
clump them together.