Fix realpath().
There were a few detail problems. First, I noticed a useless readlink() call,
caused by me not realizing that after a successful readlink() call, I know that
everything before the name of the symlink (and thus all I keep after) is a known
directory. That is also the case after processing a ./ component.

Appending a / to the resolved path is something I only have to do if the
resolved path is not empty. If it is empty, the reference tail[-1] is invalid
anyway.

Also, when removing a component from the path, the code assumed the resolved
path to not end in a /, but then also made it end in / at the end, so processing
multiple ../ in a row would fail to remove further components.

And finally, when pasting CWD and resolution result together, I must write my
own / between them, because there may not be a / I can copy from elsewhere. In
the trivial test case of resolving a path in the local directory, the resolution
ends with stack="filename" and then we get buffer="/path/to/dir", and I need to
write my own / between them to get a valid file name.