Harmonize x86 memmove variants.
I finally found my mistake: In assembler, the conditional jump over a
conditional part has to test for the *negation* of the condition. In this case,
I want to do the backwards copy if rax < rdx, so I have to do the jump if
rax >= rdx.

Now, all x86 variants use "jae" and not "jb".